somemore ugly %lu's changed back to %d.
git-svn-id: http://svn.ratbox.org/svnroot/trunk@6196 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ modules/m_svinfo.c
@@ -107,14 +107,14 @@ static void ms_svinfo(struct Client *client_p, struct Client *source_p,
   if (deltat > ConfigFileEntry.ts_max_delta)
     {
       sendto_realops_flags(FLAGS_ALL,
-       "Link %s dropped, excessive TS delta (my TS=%lu, their TS=%lu, delta=%lu)",
+       "Link %s dropped, excessive TS delta (my TS=%lu, their TS=%lu, delta=%d)",
                  get_client_name(source_p, SHOW_IP),
-                 CurrentTime, theirtime,deltat);
+                 CurrentTime, theirtime, (int)deltat);
 
       log(L_NOTICE,
-       "Link %s dropped, excessive TS delta (my TS=%lu, their TS=%lu, delta=%lu)",
+       "Link %s dropped, excessive TS delta (my TS=%lu, their TS=%lu, delta=%d)",
                  get_client_name(source_p, SHOW_IP),
-                 CurrentTime, theirtime,deltat);
+                 CurrentTime, theirtime, (int)deltat);
       exit_client(source_p, source_p, source_p, "Excessive TS delta");
       return;
     }
@@ -122,9 +122,9 @@ static void ms_svinfo(struct Client *client_p, struct Client *source_p,
   if (deltat > ConfigFileEntry.ts_warn_delta)
     { 
       sendto_realops_flags(FLAGS_ALL,
-                 "Link %s notable TS delta (my TS=%lu, their TS=%lu, delta=%lu)",
+                 "Link %s notable TS delta (my TS=%lu, their TS=%lu, delta=%d)",
 			   get_client_name(source_p, MASK_IP),
-			   CurrentTime, theirtime, deltat);
+			   CurrentTime, theirtime, (int)deltat);
     }
 }
